api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9
Pillow==10.3.0
boto3==1.34.131
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
